+++ CHANGELOG.md
@@ -29,6 +29,7 @@ All notable changes to the Wazuh app for Splunk project will be documented in th
 ### Changed 
 
 - Upgrades jQuery version from 2.1.0 to 3.5.0 [#1118](https://github.com/wazuh/wazuh-splunk/pull/1118)
+- Updated deprecated authd options [#1172](https://github.com/wazuh/wazuh-splunk/pull/1172)
 
 ### Fixed